backend/pyproject.toml
[tool.poetry]
name = "stripe-fastapi"
version = "0.1.0"
description = ""
authors = ["ZakPorat"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
stripe = "^9.12.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
